Reorganize double underscore tests
