+++ tests/baselines/reference/doubleUnderscoreEnumEmit.types
@@ -1,0 +1,43 @@
+=== tests/cases/compiler/doubleUnderscoreEnumEmit.ts ===
+enum Foo {
+>Foo : Foo
+
+    "__a" = 1,
+>1 : 1
+
+    "(Anonymous function)" = 2,
+>2 : 2
+
+    "(Anonymous class)" = 4,
+>4 : 4
+
+    "__call" = 10
+>10 : 10
+}
+namespace Foo {
+>Foo : typeof Foo
+
+    export function ___call(): number {
+>___call : () => number
+
+        return 5;
+>5 : 5
+    }
+}
+function Bar() {
+>Bar : typeof Bar
+
+    return "no";
+>"no" : "no"
+}
+namespace Bar {
+>Bar : typeof Bar
+
+    export function __call(x: number): number {
+>__call : (x: number) => number
+>x : number
+
+        return 5;
+>5 : 5
+    }
+}
