+++ tests/baselines/reference/doubleUnderscoreExportStarConflict.errors.txt
@@ -1,0 +1,15 @@
+tests/cases/compiler/index.tsx(2,1): error TS2308: Module "./b" has already exported a member named '__foo'. Consider explicitly re-exporting to resolve the ambiguity.
+
+
+==== tests/cases/compiler/index.tsx (1 errors) ====
+    export * from "./b";
+    export * from "./c";
+    ~~~~~~~~~~~~~~~~~~~~
+!!! error TS2308: Module "./b" has already exported a member named '__foo'. Consider explicitly re-exporting to resolve the ambiguity.
+    
+==== tests/cases/compiler/b.ts (0 errors) ====
+    export function __foo(): number | void {}
+    
+==== tests/cases/compiler/c.ts (0 errors) ====
+    export function __foo(): string | void {}
+
