+++ tests/baselines/reference/doubleUnderscoreLabels.js
@@ -1,0 +1,29 @@
+//// [doubleUnderscoreLabels.ts]
+function doThing() {
+    __call: while (true) {
+        aLabel: for (let i = 0; i < 10; i++) {
+            if (i === 3) {
+                break __call;
+            }
+            if (i === 5) {
+                break aLabel;
+            }
+        }
+    }
+}
+doThing();
+
+//// [doubleUnderscoreLabels.js]
+function doThing() {
+    __call: while (true) {
+        aLabel: for (var i = 0; i < 10; i++) {
+            if (i === 3) {
+                break __call;
+            }
+            if (i === 5) {
+                break aLabel;
+            }
+        }
+    }
+}
+doThing();
