+++ tests/baselines/reference/doubleUnderscoreLabels.symbols
@@ -1,0 +1,26 @@
+=== tests/cases/compiler/doubleUnderscoreLabels.ts ===
+function doThing() {
+>doThing : Symbol(doThing, Decl(doubleUnderscoreLabels.ts, 0, 0))
+
+    __call: while (true) {
+        aLabel: for (let i = 0; i < 10; i++) {
+>i : Symbol(i, Decl(doubleUnderscoreLabels.ts, 2, 24))
+>i : Symbol(i, Decl(doubleUnderscoreLabels.ts, 2, 24))
+>i : Symbol(i, Decl(doubleUnderscoreLabels.ts, 2, 24))
+
+            if (i === 3) {
+>i : Symbol(i, Decl(doubleUnderscoreLabels.ts, 2, 24))
+
+                break __call;
+            }
+            if (i === 5) {
+>i : Symbol(i, Decl(doubleUnderscoreLabels.ts, 2, 24))
+
+                break aLabel;
+            }
+        }
+    }
+}
+doThing();
+>doThing : Symbol(doThing, Decl(doubleUnderscoreLabels.ts, 0, 0))
+
