+++ tests/baselines/reference/doubleUnderscoreReactNamespace.symbols
@@ -1,0 +1,26 @@
+=== tests/cases/compiler/index.tsx ===
+declare global {
+>global : Symbol(global, Decl(index.tsx, 0, 0))
+
+    namespace JSX {
+>JSX : Symbol(JSX, Decl(index.tsx, 0, 16))
+
+        interface IntrinsicElements {
+>IntrinsicElements : Symbol(IntrinsicElements, Decl(index.tsx, 1, 19))
+
+            __foot: any;
+>__foot : Symbol(IntrinsicElements.__foot, Decl(index.tsx, 2, 37))
+        }
+    }
+    function __make (params: object): any;
+>__make : Symbol(__make, Decl(index.tsx, 5, 5))
+>params : Symbol(params, Decl(index.tsx, 6, 21))
+}
+
+
+const thing = <__foot></__foot>;
+>thing : Symbol(thing, Decl(index.tsx, 10, 5))
+>__foot : Symbol(JSX.IntrinsicElements.__foot, Decl(index.tsx, 2, 37))
+>__foot : Symbol(JSX.IntrinsicElements.__foot, Decl(index.tsx, 2, 37))
+
+export {}
