+++ tests/baselines/reference/doubleUnderscoreReactNamespace.types
@@ -1,0 +1,27 @@
+=== tests/cases/compiler/index.tsx ===
+declare global {
+>global : typeof global
+
+    namespace JSX {
+>JSX : any
+
+        interface IntrinsicElements {
+>IntrinsicElements : IntrinsicElements
+
+            __foot: any;
+>__foot : any
+        }
+    }
+    function __make (params: object): any;
+>__make : (params: object) => any
+>params : object
+}
+
+
+const thing = <__foot></__foot>;
+>thing : any
+><__foot></__foot> : any
+>__foot : any
+>__foot : any
+
+export {}
