--- tests/baselines/reference/doubleUnderscoreSymbolsAndEmit.errors.txt
@@ -1,60 +1,0 @@
-tests/cases/compiler/index.tsx(33,1): error TS2308: Module "./b" has already exported a member named '__foo'. Consider explicitly re-exporting to resolve the ambiguity.
-
-
-==== tests/cases/compiler/index.tsx (1 errors) ====
-    declare global {
-        namespace JSX {
-            interface IntrinsicElements {
-                __foot: any;
-            }
-        }
-        function __make (params: object): any;
-    }
-    
-    enum Foo {
-        "__a" = 1,
-        "(Anonymous function)" = 2,
-        "(Anonymous class)" = 4,
-        "__call" = 10
-    }
-    namespace Foo {
-        export function ___call(): number {
-            return 5;
-        }
-    }
-    function Bar() {
-        return "no";
-    }
-    namespace Bar {
-        export function __call(x: number): number {
-            return 5;
-        }
-    }
-    
-    const thing = <__foot></__foot>;
-    
-    export * from "./b";
-    export * from "./c";
-    ~~~~~~~~~~~~~~~~~~~~
-!!! error TS2308: Module "./b" has already exported a member named '__foo'. Consider explicitly re-exporting to resolve the ambiguity.
-    
-    function doThing() {
-        __call: while (true) {
-            aLabel: for (let i = 0; i < 10; i++) {
-                if (i === 3) {
-                    break __call;
-                }
-                if (i === 5) {
-                    break aLabel;
-                }
-            }
-        }
-    }
-    doThing();
-    
-==== tests/cases/compiler/b.ts (0 errors) ====
-    export function __foo(): number | void {}
-    
-==== tests/cases/compiler/c.ts (0 errors) ====
-    export function __foo(): string | void {}
-
